Changed the frame wait logic a lot. Added a Vsync option.
The wait is taken from a VBL tic now (using GraphicsQuerry) instead of 1/60th of a sec interrupt. This means,. on PAL the frame limiter will be 50, 25, 17.5 and 12.5.
Some additional hacky checks to at least enable vsync in special cases to avoid flickering. At 1VBL (50-60hz, if your 3DO can ever hit them anyway) with offscreen scalers on, it will wait for the frame update to finish before new rendering. Effectively a vsync.
The unlimited will flicker because I intentionally let it so, to be used more as a benchmark for above 50-60fps framerates (provided you stick face to a wall and screen size is ridiculously low).

Added an option to the modmenu to select the max number of visplanes you want. Initially it was 64, but you could test with less (32-40-48-56) as normally I don't see it exceeding even 32 at most levels. That saves a lot of memory, the visplanes array is allocated at start instead of being a static array. At worst case, you can select 64 from the modmenu just to be sure. But some memory could save from the release/reload resources hiccups.

Stats options will also display visplanes at current view and max visplanes hit so far. This is for testing whether at certain levels and player views we really hit a high number or it's safe enough. There is also a check if we don't exceed max visplanes that aborts floor/ceiling visplane creation and rendering. Tested with very low visplane number (4 or 8).